[ALPS07979684] Camsys: Modify buffer handle data length
[Detial]change data type from 32 bits to 64 bits

MTK-Commit-Id: 311406852b717933c09eed37b846858af48a6131

CR-Id: ALPS07979684
Feature: [Android Default] Camera Application Basic Functions

+++ drivers/media/platform/mtk-isp/camsys/isp7s/cam_legacy/mtk_cam-hsf-def.h
@@ -35,7 +35,7 @@ struct mtk_cam_hsf_info {
 	u32 enable_raw;
 	u32 cam_module;
 	u64 chunk_iova;
-	uint32_t chunk_hsfhandle;
+	u64 chunk_hsfhandle;
 };
 
 struct cq_info {
@@ -65,7 +65,7 @@ struct mtk_cam_dma_map {
 	struct dma_buf_attachment *attach;
 	struct sg_table *table;
 	dma_addr_t dma_addr;
-	uint32_t hsf_handle;
+	u64 hsf_handle;
 };
 
 struct mtk_cam_hsf_ctrl {
